requirements.txt
asgiref==3.4.1
Django==4.0.1
django-admin-tools==0.9.2
django-filter==21.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
importlib-metadata==4.10.1
Markdown==3.3.6
phonenumbers==8.12.41
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
